feat : add a file generate star schema

+++ scripts/merge_weather_data.py
@@ -8,31 +8,31 @@ DATA_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "data")
 os.makedirs(DATA_DIR, exist_ok=True)
 
 def merge_weather_data():
-    print(" Fusion des données météo...")
+    print("🔗 Fusion des données météo...")
 
     try:
         date_str = datetime.today().strftime("%Y-%m-%d")
+        start = datetime.now() - relativedelta(years=5)
+        years_range = f"{start.year}_{datetime.now().year}"
 
-        TODAY = datetime.now().date()
-        START_DATE = (datetime.now() - relativedelta(years=5)).date()
-        YEARS_RANGE = f"{START_DATE.year}_{TODAY.year}"
-
-        hist_path = os.path.join(DATA_DIR, f"openmeteo_hist_{YEARS_RANGE}.csv")
+        hist_path = os.path.join(DATA_DIR, f"openmeteo_hist_{years_range}.csv")
         recent_path = os.path.join(DATA_DIR, f"stats_weather_{date_str}.csv")
         output_path = os.path.join(DATA_DIR, "merge_weather.csv")
+
+        if not os.path.exists(hist_path) or not os.path.exists(recent_path):
+            print("❌ Un ou plusieurs fichiers de données sont manquants.")
+            return
 
         df_hist = pd.read_csv(hist_path)
         df_recent = pd.read_csv(recent_path)
 
-        df_hist["date"] = pd.to_datetime(df_hist["date"])
+        df_hist["date"] = pd.to_datetime(df_hist["date"], errors="coerce")
         df_recent = df_recent.rename(columns={"day": "date"})
-        df_recent["date"] = pd.to_datetime(df_recent["date"])
+        df_recent["date"] = pd.to_datetime(df_recent["date"], errors="coerce")
 
         merged = pd.concat([df_hist, df_recent], ignore_index=True)
         merged.to_csv(output_path, index=False)
+        print(f"✅ Fichier fusionné : {output_path}")
 
-        print(f" Fusion terminée : {output_path}")
-    except FileNotFoundError as e:
-        print(f" Fichier manquant : {e.filename}")
     except Exception as e:
-        print(f" Erreur : {e}")
+        print(f"❌ Erreur lors de la fusion : {e}")
